[CASCL-1323] chore(kubectl-datadog): split autoscaling install.go into smaller files
Pure file-level reorganization within the install package — no behavior
changes. Prepares for a future `update` subcommand that will reuse the
orchestration in steps.go (the install command is already idempotent and
most of its logic carries over).

Layout:

- install_mode.go, create_karpenter_resources.go, inference_method.go:
  pflag enum types and their String/Set/Type receiver methods (verbose
  boilerplate that obscured the install logic in the original file).
- install.go (now slim): cobra command wiring — flag-bound globals,
  options struct, complete/validate. To be duplicated by `update`.
- steps.go: orchestration — run() plus the individual steps
  (CloudFormation stacks, aws-auth ConfigMap, Helm release, NodePool
  resources, cluster-info snapshot) and the display messages. To be
  reused by `update` once run() is parameterized.

Tests are split symmetrically. Logical content (excluding comments,
blank lines, and import blocks) is byte-identical to the original.
